# openapi.yaml
openapi: 3.0.3
info:
  title: SpendBoard API
  version: 1.0.0
  description: |
    JSON API expected by the SpendBoard frontend service boundary.

    The frontend currently ships with an in-browser mock implementation, but
    these operations are the contract a real backend must implement. Monetary
    values in responses are integer cents; transaction input accepts a decimal
    string amount so the backend can apply the same validation as the form.
servers:
  - url: http://127.0.0.1:8090
    description: Local development server
tags:
  - name: Authentication
  - name: Categories
  - name: Transactions
  - name: Summary
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Sign in
      description: Creates the authenticated session and returns the signed-in user.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Signed-in user. The server also establishes the session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/logout:
    post:
      tags: [Authentication]
      summary: Sign out
      operationId: logout
      responses:
        '204':
          description: Session ended.
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/me:
    get:
      tags: [Authentication]
      summary: Get the current user
      description: Returns the signed-in user, or null when no session exists. This endpoint is public so the frontend can decide whether to redirect to the login page.
      operationId: getCurrentUser
      security: []
      responses:
        '200':
          description: The currently authenticated user, or null when unauthenticated.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/User'
                  - type: 'null'

  /categories:
    get:
      tags: [Categories]
      summary: List categories
      operationId: listCategories
      responses:
        '200':
          description: Categories sorted by name ascending.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Categories]
      summary: Create a category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '201':
          description: Created category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Category name already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /categories/{categoryId}:
    parameters:
      - $ref: '#/components/parameters/CategoryId'
    get:
      tags: [Categories]
      summary: Get a category
      operationId: getCategory
      responses:
        '200':
          description: Category details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Categories]
      summary: Rename a category
      operationId: renameCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '200':
          description: Renamed category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Category name already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Categories]
      summary: Delete a category
      description: Fails while any transaction is assigned to the category.
      operationId: deleteCategory
      responses:
        '204':
          description: Category deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Category contains transactions and cannot be deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /transactions:
    get:
      tags: [Transactions]
      summary: List transactions
      description: Returns matching transactions newest first.
      operationId: listTransactions
      parameters:
        - $ref: '#/components/parameters/CategoryIdQuery'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: Matching transactions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/InvalidDate'
    post:
      tags: [Transactions]
      summary: Create a transaction
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '201':
          description: Created transaction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Referenced category does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /transactions/{transactionId}:
    parameters:
      - $ref: '#/components/parameters/TransactionId'
    put:
      tags: [Transactions]
      summary: Replace a transaction
      operationId: updateTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '200':
          description: Updated transaction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Transactions]
      summary: Delete a transaction
      operationId: deleteTransaction
      responses:
        '204':
          description: Transaction deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /summary:
    get:
      tags: [Summary]
      summary: Get the date-range summary
      operationId: getSummary
      parameters:
        - $ref: '#/components/parameters/StartDateRequired'
        - $ref: '#/components/parameters/EndDateRequired'
      responses:
        '200':
          description: Income, expense, balance and category breakdown for the range.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Summary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/InvalidDate'

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: spendboard_session
      description: Session cookie established by POST /auth/login.
  parameters:
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema:
        type: string
    TransactionId:
      name: transactionId
      in: path
      required: true
      schema:
        type: string
    CategoryIdQuery:
      name: categoryId
      in: query
      required: false
      schema:
        type: string
    StartDate:
      name: startDate
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/ISODate'
    EndDate:
      name: endDate
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/ISODate'
    StartDateRequired:
      name: startDate
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/ISODate'
    EndDateRequired:
      name: endDate
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/ISODate'
  responses:
    BadRequest:
      description: Validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication is required or credentials are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InvalidDate:
      description: Date or date range is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    ISODate:
      type: string
      format: date
      example: '2026-01-31'
    User:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
        username:
          type: string
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    Category:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    TransactionInput:
      type: object
      required: [name, categoryId, amount, date]
      properties:
        name:
          type: string
          minLength: 1
        categoryId:
          type: string
        amount:
          type: string
          description: Decimal amount string with at most two fractional digits; must not be zero.
          pattern: '^-?\\d+([.,]\\d{1,2})?$'
          example: '-75.50'
        date:
          $ref: '#/components/schemas/ISODate'
        note:
          type: string
          default: ''
    Transaction:
      type: object
      required: [id, name, amountCents, date, note, categoryId, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        amountCents:
          type: integer
          description: Positive for income and negative for expense.
        date:
          $ref: '#/components/schemas/ISODate'
        note:
          type: string
        categoryId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CategorySummary:
      type: object
      required: [categoryId, categoryName, totalCents, count, percentage]
      properties:
        categoryId:
          type: string
        categoryName:
          type: string
        totalCents:
          type: integer
        count:
          type: integer
          minimum: 0
        percentage:
          type: number
          format: double
          minimum: 0
    Summary:
      type: object
      required: [totalIncomeCents, totalExpenseCents, netBalanceCents, income, expenses]
      properties:
        totalIncomeCents:
          type: integer
        totalExpenseCents:
          type: integer
          description: Sum of negative expense transactions.
        netBalanceCents:
          type: integer
        income:
          type: array
          items:
            $ref: '#/components/schemas/CategorySummary'
        expenses:
          type: array
          items:
            $ref: '#/components/schemas/CategorySummary'
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string

security:
  - sessionCookie: []